Reproduces a race condition where RangeLookup returns a range that does not contain a specified key
To reproduce the scenario, we will look up a range while the meta2
space is being split. We will then utilize a fact that the command
queue does not block the range lookup request if its request header
key does not overlap with the request header keys of commands
issued for the range split.

More specifically, the test takes the following steps:
1) A store receives a range split request that splits the meta2 space.
   The split key is "\000\000meta2b".
2) The replica starts processing the split request with timestamp T.
3) While the replica is updating the range descriptors, the store
   receives a range lookup request for meta2 key "\000\000meta2a" with
   timestamp T' (> T).
4) The range split completes and the range descriptors are updated.
5) The replica processes the range look up request. It scans key ranges
   ["\x00\x00meta2a\x00", end="\x00\x00meta3") and finds the end key
   of the right-side of the split range (start key = "\000\000meta2b").
6) The result range of lookup for key "\000\000meta2a" does not
   contain the key.